requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
awscli==1.18.174          # via -r requirements/base.in
botocore==1.19.14         # via awscli, s3transfer
colorama==0.4.3           # via awscli
docopt==0.6.2             # via -r requirements/base.in
docutils==0.15.2          # via awscli
edx-ccx-keys==1.1.0       # via -r requirements/base.in
edx-opaque-keys==2.1.1    # via -r requirements/base.in, edx-ccx-keys
graphitesend==0.10.0      # via -r requirements/base.in
importlib-metadata==2.0.0  # via stevedore
jmespath==0.10.0          # via botocore
path.py==12.5.0           # via -r requirements/base.in
path==15.0.0              # via path.py
pbr==5.5.1                # via -r requirements/base.in, stevedore
psutil==5.7.3             # via -r requirements/base.in
pyasn1==0.4.8             # via rsa
pymongo==3.11.0           # via -r requirements/base.in, edx-opaque-keys
python-dateutil==2.8.1    # via -r requirements/base.in, botocore
python-gnupg==0.4.6       # via -r requirements/base.in
pyyaml==5.3.1             # via -r requirements/base.in, awscli
rsa==4.5                  # via awscli
s3transfer==0.3.3         # via awscli
six==1.15.0               # via edx-ccx-keys, edx-opaque-keys, python-dateutil
stevedore==3.2.2          # via edx-opaque-keys
urllib3==1.25.11          # via botocore
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
